+++ bidscoin/Modules/_constants.py
@@ -1,0 +1,3 @@
+
+ignoremodality = '__ignore__'
+unknownmodality = '__unknown__'